main,config: set a working directory
Set and change to working directory. By default workdir is the current
directory. This can be changed with the `--workdir` flag or by setting
the config option "rundir" in the config file.

The localstatedir, rundir and logdir config options are set based on
workdir by default. To do this cleanly, the defaults are set after all
any provided config values are initialized.