Move `Context` to the `transaction` package
Makes more sense to have Context in transaction, as it was mainly
created to be used for transactions.

TODO: Rename `Context` to `TxnCtx` to avoid confusion
TODO: Rename `context` to `gdctx` to avoid confusion

`Context.{Set,Get,Delete}` work on the store now. A "transaction"
namespace is created in the store, and each transaction gets its own
transaction under it.