Don't let child process become zombie
glusterfsd, when spawned will exit because:
* Directories in logfile path and pidfile path don't exist
* It tries to use portmap and communicate with glusterd to get volfile.
  This isn't implemented yet.

When glusterfsd dies immediately after spawn, it'll become zombie if
exit status of the process isn't read by parent.

This change also moves the code that reads pid file to place where it
really should be. Also checks if process is really running before it
returns success to the caller.

Signed-off-by: Prashanth Pai <ppai@redhat.com>