Fix daemon start
The args passed to exec.Command() was a single string. It should really
be multiple strings, each one being a separate param.

The process of checking if the daemon was running or not was faulty.
This is because the child process may have forked again and the parent
(child of glusterd2) may have exited during the daemonizing procedure
performed by the child.

Signed-off-by: Prashanth Pai <ppai@redhat.com>