Fix volume ID mismatch
Problem:
The volume ID was being generated multiple times, once each on the node
or instance which make up the volume. This is because the method
'validateVolumeCreate' was run on all nodes that are part of transaction
and this method was generating a new volume ID. As a result the volume
ID in xattr set on each brick would be different.

Fix:
Only the initiator node should generate the volume id once, set it in the
transaction context, so that every other node/instance can make use of
that.

Signed-off-by: Prashanth Pai <ppai@redhat.com>