XXX: libmgmt: Rough first hack
This doesn't compile yet, but it's meant as a rough guide to get the
ball rolling here. The idea is to throw this back and forth between me
and ppai/kshlm. Note that I had to copy the vendor/* contents from mgmt
into vendor/ here.

Don't merge! Patch on top or rebase, and eventually we'll squash this
all down.