elasticetcd: set PURLs instead of CURLs when volunteering
The initial-cluster must be formed using the Peer URLs, and not the
Client URLs.

The initial-cluster is formed by the urls of the nominees, which are
inturn set to the urls set on volunteering.

Till now, I was setting the CURLs on nomination mistakenly, and somehow
this worked. I have no idea how.