Revoke liveness on graceful shutdowns
The default etcd session TTL is 30 seconds which is quite high. When a
glusterd2 instance is shutdown, it takes a while for other nodes to
notice this as TTL for the liveness key should expire.

When a glusterd2 instance is shutdown gracefully (SIGTERM), it can
simply delete its liveness key from the store and hence other nodes can
see this instantly.

Signed-off-by: Prashanth Pai <ppai@redhat.com>