Prevent redundant port allocation for a brick process.
Every time Args() method is called on a brick process object a port
is allocated to it and is returned along with other arguments
required to start the brick process. This port allocation shouldn't
happen more than once for a brick process. This commit makes sure
that we do not try to generate the argument string if it has already
been generated for a brick process thus avoiding redundant port
allocation.

Fixes: #621

Signed-off-by: Samikshan Bairagya <samikshan@gmail.com>