Remove ResponseHeaderTimeout
...that was set to 3 seconds. ResponseHeaderTimeout specifies the amount
of time to wait for a server's response headers after fully writing the
request. This time does not include the time to read the response body.

Instead, set 'Timeout' field of http.Client to 30s which includes TCP
connection time and also time taken to read the response body.

Regardless of how many nodes (even 100s) make up a volume, 30 seconds
is a large enough timeout for completion of a request.

Note:
* We can expose --timeout as a CLI option in future if deemed required.
* We may also introduce async APIs for long running tasks such as
  dynamic volume provisioning.

Signed-off-by: Prashanth Pai <ppai@redhat.com>