Set request logger to logrus.FieldLogger interface type
Request logger was eariler being set to logrus.Entry type.

With this change, request logger is set to logrus.FieldLogger which is
an interface and logrus.Entry implements this interface. Further,
renamed variables to make difference between logger and entry more
clear.

This change also fixes a crash in transaction.CreateLockFuncs() which
occurs when logger is nil. Logger will be nil if context passed is
not dervided from the REST request.

Signed-off-by: Prashanth Pai <ppai@redhat.com>