Fix GETSPEC returning false error to SHD
Clients (FUSE/gfapi) request list of gd2 servers for HA as part of the
GETSPEC request. Glusterd2 fulfils this request by getting gd2 servers
from peer ID stored in Volinfo. For normal clients, volfile ID
requested in GETSPEC request is always the name of the volume.

However, self heal daemon is also a client and uses volfile ID set to
"gluster/glustershd" in GETSPEC request. As that volfile ID is not a
volume, it needs to be handled in a special way.

Signed-off-by: Prashanth Pai <ppai@redhat.com>