Fix 100% CPU usage after peer add
When a peer gets a request to join a cluster, only the global events
listener was being shutdown, and not the entire events framework. This
resulted in the liveness watcher to continue running which became an
infinite loop because the etcd watch channel won't block.

    wch := store.Store.Watch(...)
    for {
            select {
                case resp := <-wch: // this wouldn't block
                    ...
                case <-l.stopCh:
                    return
            }
    }

This change will ensure that the entire events framework including
the liveness watcher shuts down before the store access shuts down.

Signed-off-by: Prashanth Pai <ppai@redhat.com>