Remove 'workdir' config in favour of 'localstatedir'
For all practical purposes, in real usage and in tests, the configurable
paths `workdir` and `localstatedir` are set to the same directory which
defaults to `/var/lib/glusterd2` or `/user/local/var/lib/glusterd2`

Unlike glusterd1, glusterd2 does not daemonize itself. The working
directory of the process can be meaningfully set to either current
working directory from where glusterd2 process is started or to
`localstatedir` where glusterd2 stores all its data. We choose the
latter.

This change also reduces ambiguity for users and one less thing to
configure.

Signed-off-by: Prashanth Pai <ppai@redhat.com>