Enable Rest Authentication by default
if a restauth key is missing in config
file or if restauth is set to true,
glusterd2 will do token authentication.

Authentication can be disabled by
by setting restauth key in the configuration
to false.

updated test cases to use authentication
if auth file is present

added default rest authentication in glustercli

Signed-off-by: Madhu Rajanna <mrajanna@redhat.com>