Add support for xlator options containing dots
Volume set/reset options can be of one of these forms:
<graph>.<xlator>.<option-name>
<xlator>.<option-name>

As xlator option name such as `transport.socket.ssl-enabled` can
themselves contain dots, for the options.SplitKey() function it was
hard to figure out which is graph and which is xlator.

This change ensures that graph and xlator component are properly split
for xlator options that contain dots.

Signed-off-by: Prashanth Pai <ppai@redhat.com>