e2e: stop using a gds global
A fair number of test cases were sharing the "gds" variable defined
in the volume ops test. Instead of sharing this variable across
test instances we make it local to the function and pass the
parameter as needed. The one unfortunate thing about this change
is that is requires a few wrappers around the functions passed
to t.Run to capture the local variable.

Signed-off-by: John Mulligan <jmulligan@redhat.com>