e2e: optionally start and manage etcd external to glusterd2
This change establishes the ability for the e2e tests to
start an etcd instance that all glusterd2 instances under
test will use, but that is not managed by glusterd2's
internal elastic etcd feature.

A new flag and corresponding global is established to control
whether the glusterd2 instances should use the internal,
elastic etcd, mode or if they should rely on an externally
deployed etcd instance. The flag -external-etcd is used to
control whether the testCluster instance will start up etcd
and have glusterd2 use it or not.

Note that this etcd instance is still internal to the test suite
and the tests should only require an etcd binary on the path.

Signed-off-by: John Mulligan <jmulligan@redhat.com>