middleware: log the incoming request prior to processing
It is useful to log the request as it is being made to the
server rather than relying entirely on logging generated
after the request. Logging before the request is processed can
help a person debugging with understanding when the client
first made the request.
Log the client's user agent string.

Signed-off-by: John Mulligan <jmulligan@redhat.com>