Make etcd client TLS options configurable
This change adds three new options to glusterd2's config which enables
using TLS when connecting to an external etcd cluster. The options added
are:

        --etcd-client-cert-file
        --etcd-client-key-file
        --etcd-client-ca-file

The above options are modeled after similar options present in `etcdctl`
utility.

The scope of this change is limited to external/remote etcd cluster and
doesn't alter things for elasticetcd.

Signed-off-by: Prashanth Pai <ppai@redhat.com>