glusterd2: Support setting Peer and Cluster IDs from the environment
This is an extension to b8f9625, which makes it much easier to provision
and deploy GD2 clusters in a K8S.

With this patch GD2 reads GD2_PEER_ID and GD2_CLUSTER_ID from the
environment if available, and saved into LOCALSTATEDIR/uuid.toml.
If both environment variables and the uuid file have IDs, preference is
given to the environment variables, over the uuid file.

Closes #1113