Fix TestWriteStatedump test
The unit test (TestWriteStatedump) uses a filename derived from current
timestamp and the actual code (WriteStatedump) being tested also uses
filename of the same pattern but that one may have another timestamp as
that code executes slightly later. Hence, it can so happen that the file
name written by WriteStatedump() is different than the filename expected
by the test. As these two timestamps are at seconds level granularity,
we didn't see it fail earlier.

This change modifies the test to look for a pattern that resembles
statedump file.

Signed-off-by: Prashanth Pai <ppai@redhat.com>