pmap: Let brick choose it's own port
With glusterfs change https://review.gluster.org/21249, glusterd2 can choose
to not to pass an explicit port when starting brick process. The brick process
shall choose a free port provided by the kernel and reports this port number
to glusterd2 via a SIGN IN request.

The pmap package will no longer manage and track available free ports. This
means that running multiple glusterd2 instances on same machine (e2e) are
now not prone to races in finding a free port to assign to a brick.

Signed-off-by: Prashanth Pai <ppai@redhat.com>