Tracing: Trace volume option set command.
Add instrumentation to trace volume option set transaction. Add relevant
attributes to the trace like request ID, volume name and the options along
with the values to be set on the volume. This is to help in debugging in
case the command fails for some reason. The instrumentation to trace steps
of this transaction is already in place in the common transaction and
steps framework code.

Fixes: #1299

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>