Tracing: Trace volume expand (add-brick) command.
Add instrumentation to trace volume expand (add-brick) transaction. Add
relevant attributes to the trace like request ID, volume name and the
bricks being added to the volume. This is to help in debugging in case
the command fails for some reason. The instrumentation to trace steps of
this transaction is already in place in the common transaction and steps
framework code.

Fixes: #1311

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>