nightly-container: Use 'command' instead of 'service'
I had asked for the usage of the 'service' module to enable services in
the Ansible provisioner. Forgetting that 'service' module does not work
with Ansible-Buildah as systemd isn't running in the build container
(there is even a comment about this).

This commit changes back to using the 'command' module.