Tracing: Introduce new Jaeger trace config options and management logic.
The new client configurable options that allows trace control are,

jaeger-sampler:
---------------
This option can take any one of the following values. Each value
maps to,
 - 0 (Never): Disable tracing.
 - 1 (Always): Samples all traces.
 - 2 (Probabilistic): A probabilistic sampler. Sample traces based on
   an additional option called "sampling fraction" described below using
   which Jaeger decides whether to trace the operation or not.

jaeger-sample-fraction:
-----------------------
If "jaeger-sampler-type" is set to 2 (probabilistic), then an option to
set the sampling fraction (or frequency) can be specified. By default this
value is set to 0.1 (i.e. sample every 1 in 10 traces). Valid values for
this option ranges from >0.0 and <1.0. Higher the value, the greater the
probability that a trace is sampled. This allows a control on the volume
of traces captured in a highly scaled environment.

Other changes in this commit:
- Encapsulate Jaeger specific trace configuration information in a separate
structure that is instantiated only once and updated subsequently.

- Factor out code that validates and sets trace config information into
separate functions for other modules to re-use.

closes #1368

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>