Tracing: Implement gd2 plugin that enables tracing across the cluster.
This commit implements a gd2 plugin that allows management of tracing
operations across the cluster. This change-set implements the server side
handling of the request to enable tracing on all gd2 nodes. The
pre-condition to execute this transaction is that there shouldn't be any
existing trace configuration in etcd. The steps
involved in the transaction are,
 1. Node receiving the request Validates the passed tracing configuration,
 2. Node Stores the tracing configuration in etcd, and,
 3. Set the in-memory trace configuration on all nodes

Failure in steps 2 and 3 will result in the undo logic restoring the
previous configuration both in memory and in etcd.

closes #1368

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>