Tracing: Apply trace config from store if present on GD2 start-up
If no start-up options are specified (i.e. command line or config file),
but a valid trace config exists in the store, then read the config as
part of GD2 start-up sequence and apply the trace config on the node that
is coming up. This way, all the GD2 nodes read the trace configuration
and apply it on themselves on start-up. This scenario is applicable when,
 - trace configuration was applied via glustercli (not start-up config)
   and,
 - GD2 node(s) restart for some reason. If there's a valid trace config
   in the store, then it must be applied when a node comes up.

closes #1368

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>