nightly-container: Enable lvmetad & udev for lvm
LVM config is modified to enable lvmetad and udev support. The lvm2
tmpfiles config is also removed to prevent /run/lvm bind-mounts from
being overwritten.

This requires that lvmetad is enabled on the host, and the
lvmetad.socket and udev control socket from the host are bind-mounted
within the container at /run/lvm/lvmetad.socket and /run/udev/control
respectively.