brick-mux : fix spurious socket connect failure issue during volume stop
Change the logic of using the parent socket file to send the detach
request by picking it up from /proc/<pid>/cmdline where pid is the
process id of the parent brick. During Multiplex, this would avoid
creating a hard link of a socket file.

Credits : avishwan@redhat.com for helping with the utility function

Fixes: #1468
Signed-off-by: Atin Mukherjee <amukherj@redhat.com>