Fix auto creation of block host volume
When the block host cluster options are set to values other
than default, the block volume creation fails on the first
attempt, but succeeds on the subsequent attempts. This is
due to the initialization of block vol create req before
reading the cluster options. In this patch, change the order
of the same to fix the issue.

Signed-off-by: Poornima G <pgurusid@redhat.com>